+++ Pinky/Pinky/PNKDilatedConvolutionInternalLayer.metal
@@ -8,9 +8,9 @@
 
 using namespace metal;
 
-constant ushort2 dilationRate[[function_constant(0)]];
-constant ushort2 kernelGap[[function_constant(1)]];
-constant ushort2 stride[[function_constant(2)]];
+constant uint2 dilationRate[[function_constant(0)]];
+constant uint2 kernelGap[[function_constant(1)]];
+constant uint2 stride[[function_constant(2)]];
 constant const ushort activationType [[function_constant(3)]];
 constant const bool hasAlphaBuffer [[function_constant(4)]];
 constant const bool hasBetaBuffer [[function_constant(5)]];
@@ -21,14 +21,14 @@ constant const bool hasBetaBuffer [[function_constant(5)]];
 /// loop body calculates the positions of the pixel in the input and output textures and then copies
 /// the pixel.
 template <typename U, typename V>
-void space2Patch(constant ushort2 *fullPaddingTF [[buffer(0)]], U inputTexture [[texture(0)]],
-                 V outputTexture [[texture(1)]], ushort3 gid [[thread_position_in_grid]]) {
-  const ushort2 inputTextureSize(inputTexture.get_width(), inputTexture.get_height());
-  const ushort2 outputTextureSize(outputTexture.get_width(), outputTexture.get_height());
-  const ushort2 positionInPatch = gid.xy;
-  const ushort2 leftTopShift = fullPaddingTF[0] / 2;
+void space2Patch(constant uint2 *fullPaddingTF [[buffer(0)]], U inputTexture [[texture(0)]],
+                 V outputTexture [[texture(1)]], uint3 gid [[thread_position_in_grid]]) {
+  const uint2 inputTextureSize(inputTexture.get_width(), inputTexture.get_height());
+  const uint2 outputTextureSize(outputTexture.get_width(), outputTexture.get_height());
+  const uint2 positionInPatch = gid.xy;
+  const uint2 leftTopShift = fullPaddingTF[0] / 2;
 
-  const ushort2 patchSizeWithGap = (outputTextureSize + kernelGap) / dilationRate;
+  const uint2 patchSizeWithGap = (outputTextureSize + kernelGap) / dilationRate;
 
   if (any(positionInPatch >= patchSizeWithGap)) {
     return;
@@ -36,17 +36,17 @@ void space2Patch(constant ushort2 *fullPaddingTF [[buffer(0)]], U inputTexture [
 
   const bool inPatch = all(positionInPatch < patchSizeWithGap - kernelGap);
 
-  for (ushort patchX = 0; patchX < dilationRate.x; ++patchX) {
-    for (ushort patchY = 0; patchY < dilationRate.y; ++patchY) {
-      const ushort2 patchIndex = ushort2(patchX, patchY);
-      const ushort2 writePosition = positionInPatch + patchIndex * patchSizeWithGap;
+  for (uint patchX = 0; patchX < dilationRate.x; ++patchX) {
+    for (uint patchY = 0; patchY < dilationRate.y; ++patchY) {
+      const uint2 patchIndex = uint2(patchX, patchY);
+      const uint2 writePosition = positionInPatch + patchIndex * patchSizeWithGap;
       if (any(writePosition >= outputTextureSize)) {
         continue;
       }
-      const ushort2 readPositionWithShift = patchIndex + positionInPatch * dilationRate;
+      const uint2 readPositionWithShift = patchIndex + positionInPatch * dilationRate;
       const bool readPositionIsValid = inPatch && all(readPositionWithShift >= leftTopShift) &&
           all(readPositionWithShift < inputTextureSize + leftTopShift);
-      const ushort2 readPosition = readPositionWithShift - leftTopShift;
+      const uint2 readPosition = readPositionWithShift - leftTopShift;
       const half4 pixel =
           readPositionIsValid ? static_cast<half4>(lt::read(inputTexture, readPosition, gid.z)) : 0;
       lt::write(outputTexture, pixel, writePosition, gid.z);
@@ -54,17 +54,17 @@ void space2Patch(constant ushort2 *fullPaddingTF [[buffer(0)]], U inputTexture [
   }
 }
 
-kernel void space2PatchArray(constant ushort2 *fullPaddingTF [[buffer(0)]],
+kernel void space2PatchArray(constant uint2 *fullPaddingTF [[buffer(0)]],
                              texture2d_array<half, access::read> inputTexture [[texture(0)]],
                              texture2d_array<half, access::write> outputTexture [[texture(1)]],
-                             ushort3 gid [[thread_position_in_grid]]) {
+                             uint3 gid [[thread_position_in_grid]]) {
   space2Patch(fullPaddingTF, inputTexture, outputTexture, gid);
 }
 
-kernel void space2PatchSingle(constant ushort2 *fullPaddingTF [[buffer(0)]],
+kernel void space2PatchSingle(constant uint2 *fullPaddingTF [[buffer(0)]],
                               texture2d<half, access::read> inputTexture [[texture(0)]],
                               texture2d<half, access::write> outputTexture [[texture(1)]],
-                              ushort3 gid [[thread_position_in_grid]]) {
+                              uint3 gid [[thread_position_in_grid]]) {
   space2Patch(fullPaddingTF, inputTexture, outputTexture, gid);
 }
 
@@ -75,31 +75,31 @@ kernel void space2PatchSingle(constant ushort2 *fullPaddingTF [[buffer(0)]],
 /// textures, reads it from the input, activates it and writes to the output.
 template <typename U, typename V>
 void patch2Space(constant half4 *alpha, constant half4 *beta, U inputTexture, V outputTexture,
-                 ushort3 gid) {
-  const ushort2 inputTextureSize(inputTexture.get_width(), inputTexture.get_height());
-  const ushort2 outputTextureSize(outputTexture.get_width(), outputTexture.get_height());
-  const ushort2 positionInPatch = gid.xy;
+                 uint3 gid) {
+  const uint2 inputTextureSize(inputTexture.get_width(), inputTexture.get_height());
+  const uint2 outputTextureSize(outputTexture.get_width(), outputTexture.get_height());
+  const uint2 positionInPatch = gid.xy;
 
-  const ushort2 patchSizeWithGap = (inputTextureSize + kernelGap) / dilationRate;
-  const ushort2 patchSize = patchSizeWithGap - kernelGap;
+  const uint2 patchSizeWithGap = (inputTextureSize + kernelGap) / dilationRate;
+  const uint2 patchSize = patchSizeWithGap - kernelGap;
   if (any(positionInPatch >= patchSize)) {
     return;
   }
 
-  for (ushort patchX = 0; patchX < dilationRate.x; ++patchX) {
-    for (ushort patchY = 0; patchY < dilationRate.y; ++patchY) {
-      const ushort2 patchIndex = ushort2(patchX, patchY);
-      const ushort2 writePositionWhenNoStride =
+  for (uint patchX = 0; patchX < dilationRate.x; ++patchX) {
+    for (uint patchY = 0; patchY < dilationRate.y; ++patchY) {
+      const uint2 patchIndex = uint2(patchX, patchY);
+      const uint2 writePositionWhenNoStride =
           patchIndex + positionInPatch * dilationRate;
       if (any(writePositionWhenNoStride % stride != 0)) {
         continue;
       }
-      const ushort2 writePosition = writePositionWhenNoStride / stride;
+      const uint2 writePosition = writePositionWhenNoStride / stride;
       if (any(writePosition >= outputTextureSize)) {
         continue;
       }
 
-      const ushort2 readPosition = positionInPatch + patchIndex * patchSizeWithGap;
+      const uint2 readPosition = positionInPatch + patchIndex * patchSizeWithGap;
       const half4 value = static_cast<half4>(lt::read(inputTexture, readPosition, gid.z));
       const half4 activatedValue = pnk::ActivatedValue(value, activationType, alpha, beta,
                                                        gid.z);
@@ -112,14 +112,16 @@ kernel void patch2SpaceArray(constant half4 *alpha [[buffer(0), function_constan
                              constant half4 *beta [[buffer(1), function_constant(hasBetaBuffer)]],
                              texture2d_array<half, access::read> inputTexture [[texture(0)]],
                              texture2d_array<half, access::write> outputTexture [[texture(1)]],
-                             ushort3 gid [[thread_position_in_grid]]) {
+                             uint3 gid [[thread_position_in_grid]]) {
   patch2Space(alpha, beta, inputTexture, outputTexture, gid);
 }
 
-kernel void patch2SpaceSingle(constant half4 *alpha [[buffer(0), function_constant(hasAlphaBuffer)]],
-                              constant half4 *beta [[buffer(1), function_constant(hasBetaBuffer)]],
+kernel void patch2SpaceSingle(constant half4 *alpha [[buffer(0),
+                                                      function_constant(hasAlphaBuffer)]],
+                              constant half4 *beta [[buffer(1),
+                                                     function_constant(hasBetaBuffer)]],
                               texture2d<half, access::read> inputTexture [[texture(0)]],
                               texture2d<half, access::write> outputTexture [[texture(1)]],
-                              ushort3 gid [[thread_position_in_grid]]) {
+                              uint3 gid [[thread_position_in_grid]]) {
   patch2Space(alpha, beta, inputTexture, outputTexture, gid);
 }
