+++ Pinky/Pinky/PNKUpsampling.metal
@@ -9,14 +9,14 @@
 using namespace metal;
 using namespace pnk;
 
-const constant ushort2 shiftX(1, 0);
-const constant ushort2 shiftY(0, 1);
-const constant ushort2 shiftXY(1, 1);
+const constant uint2 shiftX(1, 0);
+const constant uint2 shiftY(0, 1);
+const constant uint2 shiftXY(1, 1);
 
 constexpr sampler bilinearSampler(filter::linear);
 
 template <typename U, typename V>
-void nearestNeighbor(U inputImage, V outputImage, ushort2 gridIndex, ushort arrayIndex) {
+void nearestNeighbor(U inputImage, V outputImage, uint2 gridIndex, uint arrayIndex) {
   const ushort2 inputSize = ushort2(inputImage.get_width(), inputImage.get_height());
   if (gridIndex.x >= inputSize.x || gridIndex.y >= inputSize.y) {
     return;
@@ -24,7 +24,7 @@ void nearestNeighbor(U inputImage, V outputImage, ushort2 gridIndex, ushort arra
 
   half4 pixel = lt::read(inputImage, gridIndex, arrayIndex);
 
-  ushort2 outputCoordinates = gridIndex * 2;
+  uint2 outputCoordinates = gridIndex * 2;
 
   lt::write(outputImage, pixel, outputCoordinates, arrayIndex);
   lt::write(outputImage, pixel, outputCoordinates + shiftX, arrayIndex);
@@ -34,13 +34,13 @@ void nearestNeighbor(U inputImage, V outputImage, ushort2 gridIndex, ushort arra
 
 kernel void nearestNeighborSingle(texture2d<half, access::read> inputImage [[texture(0)]],
                                   texture2d<half, access::write> outputImage [[texture(1)]],
-                                  ushort2 gridIndex [[thread_position_in_grid]]) {
+                                  uint2 gridIndex [[thread_position_in_grid]]) {
   nearestNeighbor(inputImage, outputImage, gridIndex, 0);
 }
 
 kernel void nearestNeighborArray(texture2d_array<half, access::read> inputImage [[texture(0)]],
                                  texture2d_array<half, access::write> outputImage [[texture(1)]],
-                                 ushort3 gridIndex [[thread_position_in_grid]]) {
+                                 uint3 gridIndex [[thread_position_in_grid]]) {
   if (gridIndex.z >= outputImage.get_array_size()) {
     return;
   }
@@ -48,24 +48,24 @@ kernel void nearestNeighborArray(texture2d_array<half, access::read> inputImage 
 }
 
 template <typename U, typename V>
-void bilinear(U inputImage, V outputImage, ushort2 gridIndex, ushort arrayIndex) {
-  const ushort2 inputSize = ushort2(inputImage.get_width(), inputImage.get_height());
+void bilinear(U inputImage, V outputImage, uint2 gridIndex, uint arrayIndex) {
+  const uint2 inputSize = uint2(inputImage.get_width(), inputImage.get_height());
   if (gridIndex.x >= inputSize.x || gridIndex.y >= inputSize.y) {
     return;
   }
 
-  ushort inputCoordX0 = gridIndex.x;
-  ushort inputCoordX1 = min(inputCoordX0 + 1, inputSize.x - 1);
+  uint inputCoordX0 = gridIndex.x;
+  uint inputCoordX1 = min(inputCoordX0 + 1, inputSize.x - 1);
 
-  ushort inputCoordY0 = gridIndex.y;
-  ushort inputCoordY1 = min(inputCoordY0 + 1, inputSize.y - 1);
+  uint inputCoordY0 = gridIndex.y;
+  uint inputCoordY1 = min(inputCoordY0 + 1, inputSize.y - 1);
 
-  half4 pixel00 = lt::read(inputImage, ushort2(inputCoordX0, inputCoordY0), arrayIndex);
-  half4 pixel10 = lt::read(inputImage, ushort2(inputCoordX1, inputCoordY0), arrayIndex);
-  half4 pixel01 = lt::read(inputImage, ushort2(inputCoordX0, inputCoordY1), arrayIndex);
-  half4 pixel11 = lt::read(inputImage, ushort2(inputCoordX1, inputCoordY1), arrayIndex);
+  half4 pixel00 = lt::read(inputImage, uint2(inputCoordX0, inputCoordY0), arrayIndex);
+  half4 pixel10 = lt::read(inputImage, uint2(inputCoordX1, inputCoordY0), arrayIndex);
+  half4 pixel01 = lt::read(inputImage, uint2(inputCoordX0, inputCoordY1), arrayIndex);
+  half4 pixel11 = lt::read(inputImage, uint2(inputCoordX1, inputCoordY1), arrayIndex);
 
-  ushort2 outputCoordinates = gridIndex * 2;
+  uint2 outputCoordinates = gridIndex * 2;
 
   lt::write(outputImage, pixel00, outputCoordinates, arrayIndex);
   lt::write(outputImage, 0.5 * (pixel00 + pixel10), outputCoordinates + shiftX, arrayIndex);
@@ -76,13 +76,13 @@ void bilinear(U inputImage, V outputImage, ushort2 gridIndex, ushort arrayIndex)
 
 kernel void bilinearSingle(texture2d<half, access::read> inputImage [[texture(0)]],
                            texture2d<half, access::write> outputImage [[texture(1)]],
-                           ushort2 gridIndex [[thread_position_in_grid]]) {
+                           uint2 gridIndex [[thread_position_in_grid]]) {
   bilinear(inputImage, outputImage, gridIndex, 0);
 }
 
 kernel void bilinearArray(texture2d_array<half, access::read> inputImage [[texture(0)]],
                           texture2d_array<half, access::write> outputImage [[texture(1)]],
-                          ushort3 gridIndex [[thread_position_in_grid]]) {
+                          uint3 gridIndex [[thread_position_in_grid]]) {
   if (gridIndex.z >= outputImage.get_array_size()) {
     return;
   }
@@ -91,7 +91,7 @@ kernel void bilinearArray(texture2d_array<half, access::read> inputImage [[textu
 
 template <typename U, typename V>
 void bilinearAligned(U inputImage, V outputImage,
-                     constant SamplingCoefficients *samplingCoefficients, ushort2 gridIndex,
+                     constant SamplingCoefficients *samplingCoefficients, uint2 gridIndex,
                      ushort arrayIndex) {
   if (gridIndex.x >= outputImage.get_width() || gridIndex.y >= outputImage.get_height()) {
     return;
@@ -109,14 +109,14 @@ void bilinearAligned(U inputImage, V outputImage,
 kernel void bilinearAlignedSingle(texture2d<half, access::sample> inputImage [[texture(0)]],
                                   texture2d<half, access::write> outputImage [[texture(1)]],
                                   constant SamplingCoefficients *samplingCoefficients [[buffer(0)]],
-                                  ushort2 gridIndex [[thread_position_in_grid]]) {
+                                  uint2 gridIndex [[thread_position_in_grid]]) {
   bilinearAligned(inputImage, outputImage, samplingCoefficients, gridIndex, 0);
 }
 
 kernel void bilinearAlignedArray(texture2d_array<half, access::sample> inputImage [[texture(0)]],
                                  texture2d_array<half, access::write> outputImage [[texture(1)]],
                                  constant SamplingCoefficients *samplingCoefficients [[buffer(0)]],
-                                 ushort3 gridIndex [[thread_position_in_grid]]) {
+                                 uint3 gridIndex [[thread_position_in_grid]]) {
   if (gridIndex.z >= outputImage.get_array_size()) {
     return;
   }
